Move `useScrollLock` to global hooks folder
Update imports.
