+++ src/client/pages/Offer/Checkout/index.tsx
@@ -30,7 +30,7 @@ import { CampaignBadge } from 'components/CampaignBadge/CampaignBadge'
 import { setupQuoteCartSession } from 'containers/SessionContainer'
 import { useVariation } from 'utils/hooks/useVariation'
 import { StartDate } from 'pages/Offer/Introduction/Sidebar/StartDate'
-import { useScrollLock, VisibilityState } from 'pages/OfferNew/Checkout/hooks'
+import { useScrollLock, VisibilityState } from 'utils/hooks/useScrollLock'
 import { UpsellCard } from 'pages/OfferNew/Checkout/UpsellCard'
 import { OfferData } from 'pages/OfferNew/types'
 import { SignFailModal } from 'pages/OfferNew/Checkout/SignFailModal/SignFailModal'
