+++ src/client/pages/OfferNew/Checkout/index.tsx
@@ -25,8 +25,9 @@ import { useFeature, Features } from 'utils/hooks/useFeature'
 import { useCurrentLocale } from 'l10n/useCurrentLocale'
 import { CloseButton } from 'components/CloseButton/CloseButton'
 import { DiscountTag } from 'components/DiscountTag/DiscountTag'
+import { useScrollLock, VisibilityState } from 'utils/hooks/useScrollLock'
 import { StartDate } from '../Introduction/Sidebar/StartDate'
-import { useScrollLock, VisibilityState, useSsnError } from './hooks'
+import { useSsnError } from './hooks'
 import { Sign, SignUiState } from './Sign'
 import { SignDisclaimer } from './SignDisclaimer'
 import { CheckoutSuccessRedirect } from './CheckoutSuccessRedirect'
